status and billing changes
